FIX: BUG#1: Some additional tiles not being added to dictionary
- Fixed a bug whereby some tiles weren't being added to the dictionary.

Changed how we add additional tiles to the dictionary. Before, we added tiles to the tilemap and then looped through the tilemap to look for new tiles to add.
Now, we add each tile to the dictionary at the same time we add them to the tilemap through the new "addWorldTiles" method.